requirements.txt
streamlit==1.28.1
requests==2.31.0
pillow==10.0.1
pandas==2.1.1
plotly==5.17.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.1.78